[feat/refactor/fix]DropSystem: Refactored drop system and add owner to entity
- refactored subfield related drop
- implement a more proper shared drop items
- fix monster drop, fix the issue that animals drop nothing
- fix scene suite loading

+++ src/main/java/emu/grasscutter/game/drop/DropSystem.java
@@ -6,18 +6,22 @@ import emu.grasscutter.data.GameData;
 import emu.grasscutter.data.common.DropItemData;
 import emu.grasscutter.data.excels.DropMaterialData;
 import emu.grasscutter.data.excels.DropTableData;
+import emu.grasscutter.game.entity.EntityItem;
 import emu.grasscutter.game.entity.EntityMonster;
 import emu.grasscutter.game.entity.GameEntity;
 import emu.grasscutter.game.inventory.GameItem;
 import emu.grasscutter.game.inventory.ItemType;
 import emu.grasscutter.game.player.Player;
 import emu.grasscutter.game.props.ActionReason;
-import emu.grasscutter.scripts.data.SceneMonster;
 import emu.grasscutter.server.game.BaseGameSystem;
 import emu.grasscutter.server.game.GameServer;
 import emu.grasscutter.server.packet.send.PacketDropHintNotify;
 import emu.grasscutter.server.packet.send.PacketGadgetAutoPickDropInfoNotify;
+import emu.grasscutter.utils.Position;
 import it.unimi.dsi.fastutil.ints.Int2ObjectMap;
+import it.unimi.dsi.fastutil.ints.Int2ObjectOpenHashMap;
+import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 
 import java.util.*;
 
@@ -79,31 +83,46 @@ public final class DropSystem extends BaseGameSystem {
         return dropData.getDropId();
     }
 
-    public List<GameItem> handleDungeonRewardDrop(int dropId, boolean doubleReward) {
-        if (!dropTable.containsKey(dropId)) return List.of();
-        var dropData = dropTable.get(dropId);
-        List<GameItem> items = new ArrayList<>();
-        processDrop(dropData, doubleReward ? 2 : 1, items);
-        return items;
+    @Nullable
+    public Collection<GameItem> handleEntityDrop(int entityId, String subfieldName) {
+        var subfieldInfo = GameData.getEntityDropDataMap().get(entityId);
+        if (subfieldInfo == null) return null;
+        for (var entry : subfieldInfo.getSubfields()) {
+            if (entry.getSubfieldName().equals(subfieldName)) {
+                var dropSubfieldData = GameData.getDropSubfieldDataMap().get(entry.getSubfieldId());
+                return processDrop(dropSubfieldData.getDropId(), 1);
+            }
+        }
+        return null;
+    }
+
+    @Nullable
+    public Collection<GameItem> handleDungeonRewardDrop(int dropId, boolean doubleReward) {
+        return processDrop(dropId, doubleReward ? 2 : 1);
     }
 
     public boolean handleMonsterDrop(EntityMonster monster) {
-        int dropId;
-        int level = monster.getLevel();
-        SceneMonster sceneMonster = monster.getMetaMonster();
+        // the drop items consist of two parts
+        // first part
+        int dropId = monster.getMonsterData().getKillDropId();
+        var items = processDrop(dropId, 1);
+        if (items == null) return false;
+
+        //second part
+        var sceneMonster = monster.getMetaMonster();
         if (sceneMonster != null) {
             if (sceneMonster.drop_tag != null) {
+                int level = monster.getLevel();
                 dropId = queryDropData(sceneMonster.drop_tag, level, monsterDrop);
             } else {
                 dropId = sceneMonster.drop_id;
             }
-        } else {
-            dropId = monster.getMonsterData().getKillDropId();
+            var drop = processDrop(dropId, 1);
+            if (drop != null) items.addAll(drop);
         }
-        if (!dropTable.containsKey(dropId)) return false;
+
+        // now dropId must be non-null
         var dropData = dropTable.get(dropId);
-        List<GameItem> items = new ArrayList<>();
-        processDrop(dropData, 1, items);
         if (dropData.isFallToGround()) {
             dropItems(items, ActionReason.MonsterDie, monster, monster.getScene().getPlayers().get(0), true);
         } else {
@@ -115,10 +134,9 @@ public final class DropSystem extends BaseGameSystem {
     }
 
     public boolean handleChestDrop(int chestDropId, int dropCount, GameEntity bornFrom) {
-        if (!dropTable.containsKey(chestDropId)) return false;
+        var items = processDrop(chestDropId, dropCount);
+        if (items == null) return false;
         var dropData = dropTable.get(chestDropId);
-        List<GameItem> items = new ArrayList<>();
-        processDrop(dropData, dropCount, items);
         if (dropData.isFallToGround()) {
             dropItems(items, ActionReason.OpenChest, bornFrom, bornFrom.getWorld().getHost(), false);
         } else {
@@ -135,29 +153,36 @@ public final class DropSystem extends BaseGameSystem {
 
     public boolean handleBossChestDrop(String dropTag, Player player) {
         int dropId = queryDropData(dropTag, bossLevel[player.getWorldLevel()], chestReward);
-        if (!dropTable.containsKey(dropId)) return false;
-        var dropData = dropTable.get(dropId);
-        List<GameItem> items = new ArrayList<>();
-        processDrop(dropData, 1, items);
+        var items = processDrop(dropId, 1);
+        if (items == null) return false;
         player.getInventory().addItems(items, ActionReason.OpenWorldBossChest);
         player.sendPacket(new PacketGadgetAutoPickDropInfoNotify(items));
         return true;
     }
 
-    private void processDrop(DropTableData dropData, int count, List<GameItem> items) {
+    /**
+     * Roll the drop reward
+     *
+     * @return if there is no drop entry for the dropId, return null.
+     */
+    @Nullable
+    private Collection<GameItem> processDrop(int dropId, int count) {
+        var dropData = dropTable.get(dropId);
+        if (dropData == null) return null;
+        Int2ObjectMap<GameItem> items = new Int2ObjectOpenHashMap<>();
+        for (int i = 0; i < count; i++) processDrop(dropData, items);
+        return items.values();
+    }
+
+    // `items` is used to store drop items and pile them
+    private void processDrop(DropTableData dropData, Int2ObjectMap<GameItem> items) {
         //TODO:Not clear on the meaning of some fields,like "dropLevel".Will ignore them.
         //TODO:solve drop limits,like everydayLimit.
-        if (count > 1) {
-            for (int i = 0; i < count; i++) processDrop(dropData, 1, items);
-            return;
-        }
-        if (dropData.getRandomType() == 0) {
-            int weightSum = 0;
-            for (var i : dropData.getDropVec()) {
-                int id = i.getId();
-                if (id == 0) continue;
-                weightSum += i.getWeight();
-            }
+        if (dropData.getRandomType() == 0) { // select one from pool
+            int weightSum = dropData.getDropVec().stream()
+                .filter(i -> i.getId() != 0)
+                .mapToInt(DropItemData::getWeight)
+                .sum();
             if (weightSum == 0) return;
             int weight = rand.nextInt(weightSum);
             int sum = 0;
@@ -167,60 +192,49 @@ public final class DropSystem extends BaseGameSystem {
                 sum += i.getWeight();
                 if (weight < sum) {
                     //win the item
-                    int amount = calculateDropAmount(i) * count;
-                    if (amount <= 0) break;
-                    if (dropTable.containsKey(id)) {
-                        processDrop(dropTable.get(id), amount, items);
-                    } else {
-                        boolean flag = true;
-                        for (var j : items) {
-                            if (j.getItemId() == id) {
-                                j.setCount(j.getCount() + amount);
-                                flag = false;
-                                break;
-                            }
-                        }
-                        if (flag) items.add(new GameItem(id, amount));
-                    }
-                    break;
+                    addReward(i, items);
+                    break; // win only one pool
                 }
             }
-        } else if (dropData.getRandomType() == 1) {
+        } else if (dropData.getRandomType() == 1) { // select multiple
             for (var i : dropData.getDropVec()) {
                 int id = i.getId();
                 if (id == 0) continue;
                 if (rand.nextInt(10000) < i.getWeight()) {
-                    int amount = calculateDropAmount(i) * count;
-                    if (amount <= 0) continue;
-                    if (dropTable.containsKey(id)) {
-                        processDrop(dropTable.get(id), amount, items);
-                    } else {
-                        boolean flag = true;
-                        for (var j : items) {
-                            if (j.getItemId() == id) {
-                                j.setCount(j.getCount() + amount);
-                                flag = false;
-                                break;
-                            }
-                        }
-                        if (flag) items.add(new GameItem(id, amount));
-                    }
+                    //win the item
+                    addReward(i, items);
                 }
             }
         }
     }
 
-    private int calculateDropAmount(DropItemData i) {
+    private void addReward(DropItemData dropEntry, Int2ObjectMap<GameItem> items) {
+        var id = dropEntry.getId();
+        int amount = calculateDropAmount(dropEntry.getCountRange());
+        if (amount <= 0) return;
+        if (dropTable.containsKey(id)) {
+            processDrop(dropTable.get(id), items);
+        } else {
+            if (items.containsKey(id)) {
+                var item = items.get(id);
+                item.setCount(item.getCount() + amount);
+            } else {
+                items.put(id, new GameItem(id, amount));
+            }
+        }
+    }
+
+    private int calculateDropAmount(String countRange) {
         int amount;
-        if (i.getCountRange().contains(";")) {
-            String[] ranges = i.getCountRange().split(";");
+        if (countRange.contains(";")) {
+            String[] ranges = countRange.split(";");
             amount = rand.nextInt(Integer.parseInt(ranges[0]), Integer.parseInt(ranges[1]) + 1);
-        } else if (i.getCountRange().contains(".")) {
-            double expectAmount = Double.parseDouble(i.getCountRange());
+        } else if (countRange.contains(".")) {
+            double expectAmount = Double.parseDouble(countRange);
             amount = (int) expectAmount;
             if (rand.nextDouble() < expectAmount - amount) amount++;
         } else {
-            amount = Integer.parseInt(i.getCountRange());
+            amount = Integer.parseInt(countRange);
         }
         return amount;
     }
@@ -228,23 +242,37 @@ public final class DropSystem extends BaseGameSystem {
     /**
      * @param share Whether other players in the scene could see the drop items.
      */
-    private void dropItem(GameItem item, ActionReason reason, Player player, GameEntity bornFrom, boolean share) {
+    public void dropItem(GameItem item, ActionReason reason, @Nullable Player player, @NotNull GameEntity bornFrom, boolean share) {
         DropMaterialData drop = GameData.getDropMaterialDataMap().get(item.getItemId());
         if ((drop != null && drop.isAutoPick()) || (item.getItemData().getItemType() == ItemType.ITEM_VIRTUAL && item.getItemData().getGadgetId() == 0)) {
-            giveItem(item, reason, player, share);
+            if (player != null) giveItem(item, reason, player, share);
         } else {
-            //TODO:solve share problem
-            player.getScene().addDropEntity(item, bornFrom, player, share);
+            var scene = bornFrom.getScene();
+            var itemData = GameData.getItemDataMap().get(item.getItemId());
+            if (itemData == null) return;
+            if (itemData.isEquip()) { // equip cannot pile, drop separately
+                float range = (1.5f + (.05f * item.getCount()));
+                for (int j = 0; j < item.getCount(); j++) {
+                    Position pos = bornFrom.getPosition().nearby2d(range).addY(0.5f);
+                    EntityItem entity = new EntityItem(scene, player, itemData, pos, 1, share);
+                    scene.addEntity(entity);
+                }
+            } else {
+                EntityItem entity = new EntityItem(scene, player, itemData,
+                    bornFrom.getPosition().nearby2d(1f).addY(0.5f), // nearby2d() will clone the position
+                    item.getCount(), share);
+                scene.addEntity(entity);
+            }
         }
     }
 
-    private void dropItems(List<GameItem> items, ActionReason reason, GameEntity bornFrom, Player player, boolean share) {
+    public void dropItems(Collection<GameItem> items, ActionReason reason, GameEntity bornFrom, Player player, boolean share) {
         for (var i : items) {
-            dropItem(i,reason,player,bornFrom,share);
+            dropItem(i, reason, player, bornFrom, share);
         }
     }
 
-    private void giveItem(GameItem item, ActionReason reason, Player player, boolean share) {
+    private void giveItem(GameItem item, ActionReason reason, @NotNull Player player, boolean share) {
         if (share) {
             for (var p : player.getScene().getPlayers()) {
                 p.getInventory().addItem(item, reason);
